+++ framework/packages/abstract-interface/src/native/ans_host.rs
@@ -1,22 +1,26 @@
-#![allow(unused)]
-#[allow(unused)]
 pub use abstract_core::ans_host::ExecuteMsgFns;
 use abstract_core::{
     ans_host::*,
     objects::{
-        pool_id::UncheckedPoolAddress, PoolMetadata, UncheckedChannelEntry, UncheckedContractEntry,
+        pool_metadata::ResolvedPoolMetadata, AnsAsset, AnsEntryConvertor, AssetEntry, ChannelEntry,
+        ContractEntry, DexAssetPairing, LpToken, PoolMetadata, PoolReference, UniquePoolId,
     },
     ANS_HOST,
 };
 use cosmwasm_std::Addr;
-use cw_asset::AssetInfoUnchecked;
 
-use cw_orch::{interface, prelude::*, state::ChainState};
-use serde_json::from_reader;
-use std::{cmp::min, collections::HashSet, env, fs::File};
+use cw_asset::{Asset, AssetInfo};
+use cw_orch::{interface, prelude::*};
+use std::env;
 
 #[interface(InstantiateMsg, ExecuteMsg, QueryMsg, MigrateMsg)]
 pub struct AnsHost<Chain>;
+
+impl<Chain: CwEnv> AnsHost<Chain> {
+    pub fn resolve<R: ClientResolve<Chain>>(&self, item: &R) -> Result<R::Output, CwOrchError> {
+        item.resolve(self)
+    }
+}
 
 impl<Chain: CwEnv> Uploadable for AnsHost<Chain> {
     #[cfg(feature = "integration")]
@@ -46,206 +50,137 @@ where
     }
 }
 
-/// Implementation for the daemon, which maintains actual state
-#[cfg(feature = "daemon")]
-use cw_orch::prelude::Daemon;
+pub trait ClientResolve<Chain: CwEnv> {
+    /// Result of resolving an entry.
+    type Output;
+    /// Resolve an entry into its value.
+    fn resolve(&self, ans_host: &AnsHost<Chain>) -> Result<Self::Output, CwOrchError>;
+}
 
-#[cfg(feature = "daemon")]
-impl AnsHost<Daemon> {
-    pub fn update_all_local(&self) -> Result<(), crate::AbstractInterfaceError> {
-        self.update_assets_local()?;
-        self.update_contracts_local()?;
-        self.update_pools_local()?;
-        Ok(())
-    }
+// cw-multi-test doesn't support raw queries, so we will have to do smart queries instead
 
-    pub fn update_assets_local(&self) -> Result<(), crate::AbstractInterfaceError> {
-        let path = env::var("ANS_HOST_ASSETS").unwrap();
-        let file =
-            File::open(&path).unwrap_or_else(|_| panic!("file should be present at {}", &path));
-        let json: serde_json::Value = from_reader(file)?;
-        let chain_name = &self.get_chain().state().chain_data.chain_name;
-        let chain_id = self.get_chain().state().chain_data.chain_id.to_string();
-        let maybe_assets = json
-            .get(chain_name)
-            .unwrap()
-            .get(chain_id)
-            .ok_or_else(|| CwOrchError::StdErr("network not found".into()))?;
+impl<Chain: CwEnv> ClientResolve<Chain> for AssetEntry {
+    type Output = AssetInfo;
 
-        /*
-
-        [[
-           "junox",
-           {
-             "native": "ujunox"
-           }
-         ],
-         [
-           "crab",
-           {
-             "cw20": "juno12wwfatmpqnxdqk0ka8rpd4ud4frc97srtv55mmlf7xunux6led8s8wtgx2"
-           }
-         ]]
-            */
-
-        let assets = maybe_assets.as_array().unwrap();
-
-        let assets_to_add: Vec<(String, AssetInfoUnchecked)> = assets
-            .iter()
-            .map(|value| {
-                let asset: (String, AssetInfoUnchecked) =
-                    serde_json::from_value(value.clone()).unwrap();
-                asset
-            })
-            .collect();
-
-        self.execute_chunked(&assets_to_add, 25, |chunk| {
-            ExecuteMsg::UpdateAssetAddresses {
-                to_add: chunk.to_vec(),
-                to_remove: vec![],
-            }
+    fn resolve(&self, ans_host: &AnsHost<Chain>) -> Result<Self::Output, CwOrchError> {
+        let mut assets: AssetsResponse = ans_host.query(&QueryMsg::Assets {
+            names: vec![self.to_string()],
         })?;
-
-        Ok(())
+        Ok(assets.assets.pop().unwrap().1)
     }
+}
 
-    pub fn update_channels_local(&self) -> Result<(), crate::AbstractInterfaceError> {
-        let path = env::var("ANS_HOST_CHANNELS").unwrap();
-        let file =
-            File::open(&path).unwrap_or_else(|_| panic!("file should be present at {}", &path));
-        let json: serde_json::Value = from_reader(file)?;
-        let chain_name = &self.get_chain().state().chain_data.chain_name;
-        let chain_id = self.get_chain().state().chain_data.chain_id.to_string();
-        let channels = json
-            .get(chain_name)
-            .unwrap()
-            .get(chain_id)
-            .ok_or_else(|| CwOrchError::StdErr("network not found".into()))?;
+impl<Chain: CwEnv> ClientResolve<Chain> for LpToken {
+    type Output = AssetInfo;
 
-        let channels = channels.as_object().unwrap();
-        let channels_to_add: Vec<(UncheckedChannelEntry, String)> = channels
-            .iter()
-            .map(|(name, value)| {
-                let id = value.as_str().unwrap().to_owned();
-                let key = UncheckedChannelEntry::try_from(name.clone()).unwrap();
-                (key, id)
-            })
-            .collect();
+    fn resolve(&self, ans_host: &AnsHost<Chain>) -> Result<Self::Output, CwOrchError> {
+        let asset_entry = AnsEntryConvertor::new(self.clone()).asset_entry();
+        asset_entry.resolve(ans_host)
+    }
+}
 
-        self.execute_chunked(&channels_to_add, 25, |chunk| ExecuteMsg::UpdateChannels {
-            to_add: chunk.to_vec(),
-            to_remove: vec![],
+impl<Chain: CwEnv> ClientResolve<Chain> for ContractEntry {
+    type Output = Addr;
+
+    fn resolve(&self, ans_host: &AnsHost<Chain>) -> Result<Self::Output, CwOrchError> {
+        let mut contracts: ContractsResponse = ans_host.query(&QueryMsg::Contracts {
+            entries: vec![self.clone()],
         })?;
-
-        Ok(())
+        Ok(contracts.contracts.pop().unwrap().1)
     }
+}
 
-    pub fn update_contracts_local(&self) -> Result<(), crate::AbstractInterfaceError> {
-        let path = env::var("ANS_HOST_CONTRACTS").unwrap();
+impl<Chain: CwEnv> ClientResolve<Chain> for ChannelEntry {
+    type Output = String;
 
-        let file =
-            File::open(&path).unwrap_or_else(|_| panic!("file should be present at {}", &path));
-        let json: serde_json::Value = from_reader(file)?;
-        let chain_name = &self.get_chain().state().chain_data.chain_name;
-        let chain_id = self.get_chain().state().chain_data.chain_id.to_string();
-        let contracts = json
-            .get(chain_name)
-            .unwrap()
-            .get(chain_id)
-            .ok_or_else(|| CwOrchError::StdErr("network not found".into()))?;
-
-        /*
-          [
-        [
-          {
-            "protocol": "junoswap",
-            "contract": "staking/crab,junox"
-          },
-          "juno1vhxnvu0zh6p707auf0ltq6scse3d2fxvp4804s54q45z29vtjleqghne5g"
-        ]
-        ]
-           */
-
-        let contracts = contracts.as_array().unwrap();
-        let contracts_to_add: Vec<(UncheckedContractEntry, String)> = contracts
-            .iter()
-            .map(|value| {
-                let contract: (UncheckedContractEntry, String) =
-                    serde_json::from_value(value.clone()).unwrap();
-                contract
-            })
-            .collect();
-
-        self.execute_chunked(&contracts_to_add, 25, |chunk| {
-            ExecuteMsg::UpdateContractAddresses {
-                to_add: chunk.to_vec(),
-                to_remove: vec![],
-            }
+    fn resolve(&self, ans_host: &AnsHost<Chain>) -> Result<Self::Output, CwOrchError> {
+        let mut channels: ChannelsResponse = ans_host.query(&QueryMsg::Channels {
+            entries: vec![self.clone()],
         })?;
-
-        Ok(())
+        Ok(channels.channels.pop().unwrap().1)
     }
+}
 
-    pub fn update_pools_local(&self) -> Result<(), crate::AbstractInterfaceError> {
-        let path = env::var("ANS_HOST_POOLS").unwrap();
-        let file =
-            File::open(&path).unwrap_or_else(|_| panic!("file should be present at {}", &path));
-        let json: serde_json::Value = from_reader(file)?;
-        let chain_name = &self.get_chain().state().chain_data.chain_name;
-        let chain_id = self.get_chain().state().chain_data.chain_id.to_string();
-        let pools = json
-            .get(chain_name)
-            .unwrap()
-            .get(chain_id)
-            .ok_or_else(|| CwOrchError::StdErr("network not found".into()))?;
+impl<Chain: CwEnv> ClientResolve<Chain> for DexAssetPairing {
+    type Output = Vec<PoolReference>;
 
-        let mut dexes_to_register: HashSet<String> = HashSet::new();
-
-        let pools = pools.as_array().unwrap();
-        let pools_to_add: Vec<(UncheckedPoolAddress, PoolMetadata)> = pools
-            .iter()
-            .map(|value| {
-                let pool: (UncheckedPoolAddress, PoolMetadata) =
-                    serde_json::from_value(value.clone()).unwrap();
-
-                dexes_to_register.insert(pool.1.dex.clone());
-
-                pool
-            })
-            .collect();
-
-        // Register the dexes
-        self.0.execute(
-            &ExecuteMsg::UpdateDexes {
-                to_add: Vec::from_iter(dexes_to_register),
-                to_remove: vec![],
-            },
-            None,
-        )?;
-
-        self.execute_chunked(&pools_to_add, 25, |chunk| ExecuteMsg::UpdatePools {
-            to_add: chunk.to_vec(),
-            to_remove: vec![],
-        })?;
-
-        Ok(())
+    fn resolve(&self, ans_host: &AnsHost<Chain>) -> Result<Self::Output, CwOrchError> {
+        let mut pool_address_list: PoolAddressListResponse =
+            ans_host.query(&QueryMsg::PoolList {
+                filter: Some(AssetPairingFilter {
+                    asset_pair: Some((self.asset_x().clone(), self.asset_y().clone())),
+                    dex: Some(self.dex().to_owned()),
+                }),
+                start_after: None,
+                limit: None,
+            })?;
+        let found = pool_address_list
+            .pools
+            .pop()
+            .ok_or(CwOrchError::StdErr(format!(
+                "Pool reference for {self} not found"
+            )))?;
+        Ok(found.1.clone())
     }
+}
 
-    pub fn execute_chunked<T, MsgBuilder>(
-        &self,
-        items: &[T],
-        chunk_size: usize,
-        mut msg_builder: MsgBuilder,
-    ) -> Result<(), crate::AbstractInterfaceError>
-    where
-        MsgBuilder: FnMut(&[T]) -> ExecuteMsg,
-    {
-        let mut i = 0;
-        while i < items.len() {
-            let chunk = &items[i..min(i + chunk_size, items.len())];
-            i += chunk.len();
-            self.0.execute(&msg_builder(chunk), None)?;
-        }
-        Ok(())
+impl<Chain: CwEnv> ClientResolve<Chain> for UniquePoolId {
+    type Output = PoolMetadata;
+
+    fn resolve(&self, ans_host: &AnsHost<Chain>) -> Result<Self::Output, CwOrchError> {
+        let mut pool_metadatas: PoolMetadatasResponse =
+            ans_host.query(&QueryMsg::PoolMetadatas { ids: vec![*self] })?;
+        Ok(pool_metadatas.metadatas.pop().unwrap().1)
+    }
+}
+
+impl<Chain: CwEnv> ClientResolve<Chain> for AnsAsset {
+    type Output = Asset;
+
+    fn resolve(&self, ans_host: &AnsHost<Chain>) -> Result<Self::Output, CwOrchError> {
+        Ok(Asset::new(self.name.resolve(ans_host)?, self.amount))
+    }
+}
+
+impl<Chain: CwEnv> ClientResolve<Chain> for AssetInfo {
+    type Output = AssetEntry;
+
+    fn resolve(&self, ans_host: &AnsHost<Chain>) -> Result<Self::Output, CwOrchError> {
+        let mut assets: AssetsResponse = ans_host.query(&QueryMsg::AssetInfos { infos: vec![] })?;
+        Ok(assets.assets.pop().unwrap().0)
+    }
+}
+
+impl<Chain: CwEnv> ClientResolve<Chain> for Asset {
+    type Output = AnsAsset;
+
+    fn resolve(&self, ans_host: &AnsHost<Chain>) -> Result<Self::Output, CwOrchError> {
+        Ok(AnsAsset {
+            name: self.info.resolve(ans_host)?,
+            amount: self.amount,
+        })
+    }
+}
+
+impl<Chain: CwEnv> ClientResolve<Chain> for PoolMetadata {
+    type Output = ResolvedPoolMetadata;
+
+    fn resolve(&self, ans_host: &AnsHost<Chain>) -> Result<Self::Output, CwOrchError> {
+        Ok(ResolvedPoolMetadata {
+            assets: self.assets.resolve(ans_host)?,
+            dex: self.dex.clone(),
+            pool_type: self.pool_type,
+        })
+    }
+}
+
+impl<Chain: CwEnv, T> ClientResolve<Chain> for Vec<T>
+where
+    T: ClientResolve<Chain>,
+{
+    type Output = Vec<T::Output>;
+
+    fn resolve(&self, ans_host: &AnsHost<Chain>) -> Result<Self::Output, CwOrchError> {
+        self.iter().map(|entry| entry.resolve(ans_host)).collect()
     }
 }